app/requirements.txt
aiodns==3.1.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
ccxt==4.2.21
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.1
fastapi==0.109.0
fastapi-cache2==0.2.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
loguru==0.7.2
multidict==6.0.4
pendulum==3.0.0
psycopg==3.1.17
psycopg-binary==3.1.17
pycares==4.4.0
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
time-machine==2.13.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.27.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
